requirements.txt
telethon
gspread
oauth2client
python-dotenv